% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{K Nearest Neighbor Algorithm}
\usage{
knn(x.mat, y.vec, testx.vec, max.neighbors)
}
\arguments{
\item{x.mat}{numeric imput feature matrix [n x p]}

\item{y.vec}{numberic input label vector [n], 
either all 0/1 for binary classification or other real numbers for regression}

\item{testx.vec}{numberic test festure vector [p], that is used to test the different hyperparameters for k (1 to MAX.K) of KNN}

\item{max.neighbors}{scalar integer, max number of neighbors}
}
\value{
numeric vector of size max.neighbors, predictions from 1 to max.neighbors
}
\description{
Wraps around c++ code to call it using r
}
\examples{
data(zip.train, packages= "ElemStatLearn")
i01 <- which(zip.train[,1] \%in\% c(0,1))
train.i <- i01[1:5]
test.i <- i01[6]
x <- zip.train[train.i,-1]
y <- zip.train[train.i, 1]
testx <- zip.train[test.i, -1]
knn(x,y,testx,3)
zip.train[test.i,1]

}
